+++ package.json
@@ -10,11 +10,10 @@
   "author": "Telk",
   "license": "MIT",
   "dependencies": {
-    "active-win": "^5.0.0",
     "electron-settings": "^3.2.0",
-    "form-data": "^3.0.0",
     "node-fetch": "^2.6.1",
-    "shelljs": "^0.8.3"
+    "active-win": "^5.0.0",
+    "form-data": "^3.0.0"
   },
   "devDependencies": {
     "electron": "^7.1.2",
